+++ chatsky/llm/langchain_context.py
@@ -3,6 +3,8 @@ LLM Utils.
 ----------
 The Utils module contains functions for converting Chatsky's objects to an LLM_API and langchain compatible versions.
 """
+
+import re
 import logging
 from typing import Literal, Union
 import asyncio
@@ -10,6 +12,7 @@ import asyncio
 from chatsky.core import Context, Message
 from chatsky.llm._langchain_imports import HumanMessage, SystemMessage, AIMessage, check_langchain_available
 from chatsky.llm.filters import BaseHistoryFilter, Return
+from chatsky.llm.prompt import Prompt, PositionConfig
 
 
 logger = logging.getLogger(__name__)
@@ -78,3 +81,87 @@ async def context_to_history(
             history.append(await message_to_langchain(response, ctx=ctx, source="ai", max_size=max_size))
 
     return history
+
+
+async def get_langchain_context(
+    system_prompt: Message,
+    ctx: Context,
+    call_prompt: Prompt,
+    prompt_misc_filter: str = r"prompt",  # r"prompt" -> extract misc prompts
+    position_config: PositionConfig = PositionConfig(),
+    **history_args,
+) -> list[Union[HumanMessage, AIMessage, SystemMessage]]:
+    """
+    Get a list of Langchain messages using the context and prompts.
+    :param system_prompt: System message to be included in the context.
+    :param ctx: Current dialog context.
+    :param call_prompt: Prompt to be used for the current call.
+    :param prompt_misc_filter: Regex pattern to filter miscellaneous prompts from context.
+        Defaults to r"prompt".
+    :param position_config: Configuration for positioning different parts of the context.
+        Defaults to default PositionConfig().
+    :param history_args: Additional arguments to be passed to context_to_history function.
+    :return: List of Langchain message objects ordered by their position values.
+    """
+    check_langchain_available()
+    logger.debug(f"History args: {history_args}")
+
+    history = await context_to_history(ctx, **history_args)
+    logger.debug(f"Position config: {position_config}")
+    prompts: list[tuple[list[Union[HumanMessage, AIMessage, SystemMessage]], float]] = []
+
+    # Add system prompt
+    if system_prompt.text != "":
+        system_prompt_obj = Prompt(message=system_prompt)
+        system_messages = await system_prompt_obj.to_langchain_messages(ctx, source="system", position_config=position_config)
+        prompts.append((system_messages, position_config.system_prompt))
+
+    # Add history
+    prompts.append((history, position_config.history))
+    logger.debug(f"System prompt: {prompts[0]}")
+
+    # Add miscellaneous prompts
+    for element_name, element in ctx.current_node.misc.items():
+        if re.compile(prompt_misc_filter).match(element_name):
+            prompt = Prompt.model_validate(element)
+            prompt_messages = await prompt.to_langchain_messages(ctx, source="human", position_config=position_config)
+            prompts.append(
+                (
+                    prompt_messages,
+                    prompt.position if prompt.position is not None else position_config.misc_prompt,
+                )
+            )
+
+    # Add call prompt
+    call_messages = await call_prompt.to_langchain_messages(ctx, source="human", position_config=position_config)
+    if call_messages:  # Only add if there are messages (non-empty text)
+        prompts.append(
+            (
+                call_messages,
+                call_prompt.position if call_prompt.position is not None else position_config.call_prompt,
+            )
+        )
+
+    # Add last turn messages
+    last_turn_request = await ctx.requests.get(ctx.current_turn_id)
+    last_turn_response = await ctx.responses.get(ctx.current_turn_id)
+
+    if last_turn_request:
+        request_prompt = Prompt(message=last_turn_request)
+        request_messages = await request_prompt.to_langchain_messages(ctx, source="human", position_config=position_config)
+        prompts.append((request_messages, position_config.last_turn))
+
+    if last_turn_response:
+        response_prompt = Prompt(message=last_turn_response)
+        response_messages = await response_prompt.to_langchain_messages(ctx, source="ai", position_config=position_config)
+        prompts.append((response_messages, position_config.last_turn))
+
+    logger.debug(f"Prompts: {prompts}")
+    prompts = sorted(prompts, key=lambda x: x[1])
+
+    # flatten prompts list
+    langchain_context = []
+    for message_block in prompts:
+        langchain_context.extend(message_block[0])
+
+    return langchain_context
